+++ src/lib/sharedConstants.js
@@ -59,9 +59,10 @@ module.exports = {
         CLI: '/tm/cli/global-settings',
         DagGlobals: '/tm/net/dag-globals',
         TrafficControl: '/tm/ltm/global-settings/traffic-control',
+        TrafficGroup: '/tm/cm/traffic-group',
         HTTPD: '/tm/sys/httpd',
         SSHD: '/tm/sys/sshd',
-        Phonehome: '/tm/sys/software/update'
+        SoftwareUpdate: '/tm/sys/software/update'
     },
     STATUS: {
         STATUS_OK: 'OK',