+++ test/unit/src/lib/declarationHandlerTests.js
@@ -321,7 +321,26 @@ describe('declarationHandler', () => {
 
         it('should send TEEM report', () => {
             const newDeclaration = {
-                name: 'new'
+                name: 'new',
+                foo: {
+                    class: 'bar'
+                },
+                controls: {
+                    class: 'Controls',
+                    userAgent: 'test userAgent'
+                },
+                Common: {
+                    class: 'Tenant',
+                    myAuth: {
+                        class: 'Authentication',
+                        radius: {
+                            serviceType: 'call-check'
+                        },
+                        ldap: {
+                            port: 654
+                        }
+                    }
+                }
             };
             const state = {
                 currentConfig: {
@@ -343,11 +362,22 @@ describe('declarationHandler', () => {
             const isAddRegKeyCalled = sinon.spy(TeemRecord.prototype, 'addRegKey');
             const isAddProvisionedModulesCalled = sinon.spy(TeemRecord.prototype, 'addProvisionedModules');
             const isCalculateAssetIdCalled = sinon.spy(TeemRecord.prototype, 'calculateAssetId');
+            const isAddJsonObjectCalled = sinon.spy(TeemRecord.prototype, 'addJsonObject');
+
+            // report should not be called
+            const isReportCalled = sinon.stub(teemDevice, 'report').rejects();
+
+            // check the record sent to reportRecord
+            let record;
+            sinon.stub(teemDevice, 'reportRecord').callsFake((recordIn) => {
+                record = recordIn;
+            });
 
             const declarationHandler = new DeclarationHandler(bigIpMock);
             declarationHandler.teemDevice = teemDevice;
             return declarationHandler.process(newDeclaration, state)
                 .then(() => {
+                    // Check that each class was called
                     assert.strictEqual(isAddClassCountCalled.calledOnce, true,
                         'should call addClassCount() once');
                     assert.strictEqual(isAddPlatformInfoCalled.calledOnce, true,
@@ -358,6 +388,101 @@ describe('declarationHandler', () => {
                         'should call addProvisionedModules() once');
                     assert.strictEqual(isCalculateAssetIdCalled.calledOnce, true,
                         'should call calculateAssetId() once');
+                    assert.strictEqual(isAddJsonObjectCalled.calledOnce, true,
+                        'should call addJsonObject() once');
+                    assert.strictEqual(isReportCalled.called, false,
+                        'report() should not have been called');
+
+                    // Check that the record body object was filled with input
+                    assert.deepStrictEqual(
+                        record.recordBody,
+                        {
+                            authenticationType: {
+                                radius: 1,
+                                tacacs: 0,
+                                ldap: 1
+                            },
+                            Authentication: 1,
+                            bar: 1,
+                            Controls: 1,
+                            Tenant: 1,
+                            modules: {},
+                            nicConfiguration: 'unknown',
+                            platform: 'unknown',
+                            platformID: 'unknown',
+                            platformVersion: 'unknown',
+                            regkey: 'unknown',
+                            userAgent: 'test userAgent'
+                        }
+                    );
+                });
+        });
+
+        it('should count complicated authentication declaration', () => {
+            const newDeclaration = {
+                name: 'new',
+                Common: {
+                    class: 'Tenant',
+                    myAuth: {
+                        class: 'Authentication',
+                        radius: { serviceType: 'call-check' },
+                        ldap: { port: 654 }
+                    },
+                    funky: {
+                        class: 'Authentication',
+                        radius: { serviceType: 'call-check' },
+                        monkey: {
+                            // Should not count these two
+                            ldap: { port: 654 },
+                            tacacs: { foo: 'bar' }
+                        }
+                    }
+                }
+            };
+            const state = {
+                currentConfig: {
+                    name: 'current'
+                },
+                originalConfig: {
+                    Common: {}
+                }
+            };
+
+            const assetInfo = {
+                name: 'Declarative Onboarding',
+                version: '1.2.3'
+            };
+            const teemDevice = new TeemDevice(assetInfo);
+
+            // check the record sent to reportRecord
+            let record;
+            sinon.stub(teemDevice, 'reportRecord').callsFake((recordIn) => {
+                record = recordIn;
+            });
+
+            const declarationHandler = new DeclarationHandler(bigIpMock);
+            declarationHandler.teemDevice = teemDevice;
+            return declarationHandler.process(newDeclaration, state)
+                .then(() => {
+                    // Check that the record body object was filled with input
+                    assert.deepStrictEqual(
+                        record.recordBody,
+                        {
+                            authenticationType: {
+                                ldap: 1,
+                                radius: 2,
+                                tacacs: 0
+                            },
+                            Authentication: 2,
+                            Tenant: 1,
+                            modules: {},
+                            nicConfiguration: 'unknown',
+                            platform: 'unknown',
+                            platformID: 'unknown',
+                            platformVersion: 'unknown',
+                            regkey: 'unknown'
+                        }
+                    );
                 });
         });
 
@@ -380,7 +505,7 @@ describe('declarationHandler', () => {
             };
             const teemDevice = new TeemDevice(assetInfo);
 
-            sinon.stub(teemDevice, 'report').rejects();
+            sinon.stub(teemDevice, 'reportRecord').rejects();
             const declarationHandler = new DeclarationHandler(bigIpMock);
             declarationHandler.teemDevice = teemDevice;
             return declarationHandler.process(newDeclaration, state);