+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes to this project are documented in this file. More detail and links can b
 - AUTOTOOL-1445: Add authentication type to DO TEEM telemetry
 - AUTOTOOL-1236: Add support for System autoCheck setting
 - AUTOTOOL-1248 ([GitHub Issue 120](https://github.com/F5Networks/f5-declarative-onboarding/issues/120)): Add support for System tmshAuditLog and guiAuditLog
+- AUTOTOOL-1322: ([GitHub Issue 96](https://github.com/F5Networks/f5-declarative-onboarding/issues/96)): Support generated MAC Masquerade on Traffic Groups according to https://support.f5.com/csp/article/K3523.   
 ### Fixed
 - AUTOTOOL-1491 ([GitHub Issue 138](https://github.com/F5Networks/f5-declarative-onboarding/issues/138)): Cannot create a device group with AFM provisioned
 - AUTOTOOL-1469: Problems with latest Azure image